agentic-day1/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-google-genai
